ML/requirements.txt
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn>=0.23.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
matplotlib>=3.8.0
pandas>=2.1.0
torchcam>=0.4.0
